+++ et-common-libs/include/transports/mm_cm_iface/message_types.h
@@ -170,7 +170,8 @@ typedef enum {
     /* Compute FW specific messages start from here */
     CM_TO_MM_MESSAGE_ID_FW_SHIRE_READY,
     CM_TO_MM_MESSAGE_ID_FW_EXCEPTION,
-    CM_TO_MM_MESSAGE_ID_FW_ERROR
+    CM_TO_MM_MESSAGE_ID_FW_ERROR,
+    CM_TO_MM_MESSAGE_ID_FW_TRACE_BUFFER_FULL
 } cm_to_mm_message_id_e;
 
 /* Status values for kernel completion message */
@@ -208,6 +209,15 @@ typedef struct {
 } __attribute__((packed, aligned(64))) cm_to_mm_message_fw_error_t;
 
 ASSERT_CACHE_LINE_CONSTRAINTS(cm_to_mm_message_fw_error_t);
+
+typedef struct {
+    cm_iface_message_header_t header;
+    uint32_t data_size;
+    uint8_t buffer_type;
+    uint8_t pad[3]; /* Padding to make struct 64-bit aligned */
+} __attribute__((packed, aligned(64))) cm_to_mm_message_fw_trace_buffer_full_t;
+
+ASSERT_CACHE_LINE_CONSTRAINTS(cm_to_mm_message_fw_trace_buffer_full_t);
 
 typedef struct {
     cm_iface_message_header_t header;